feat(cli): make base node address configurable (#34)
Description
---
Base node (where p2pool connects to) was not able to configure (it's
address) from CLI, so it must be.

Motivation and Context
---

How Has This Been Tested?
---
Starting p2pool with `--base-node-address="http://127.0.0.1:9999"` and
see that it could not connect as it tries to use new address. When set
back to `--base-node-address="http://127.0.0.1:18142"` it was working
fine again.

What process can a PR reviewer use to test or verify this change?
---
See test.


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify