feat(coinbase-extra): add coinbase extra field (#67)
Description
---
Coinbase extra field is cached now (locally) and use other miners'
coinbase extra when possible from share chain for their shares.

Motivation and Context
---
Coinbase extra field was not supported to be passed to p2pool before, so
in order to let miners pass any extra data they want, we need to be able
to do that.

Specs for the format of p2pool coinbase extra field:
https://github.com/tari-project/sha-p2pool/issues/56

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - Please specify - P2pool should be updated at all clients