feat: change peer propegation (#312)
Description
---
This changes how p2pool does peer selection

Currently, we just try to select 8 outbound peers. We try to choose
these by latency. This can cause the peers tp form islands.

This changes to peer selection to be beneficial for the network. We
choose 2 peers that are not of our squad so that we can propagate gossip
messages across the entire network, like network changes. We change 1 of
these periodically.

We still try and keep 8 peers of our squad as outbound connections, but
we drop the middle connection in terms of block height so that low nodes
can sync from us and try catcup, and we replace those 2 with 2 new
peers. This is to stop nodes from just connecting to the same peers, and
cycle some of them to ensure we have a fully connected network.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced network connectivity by introducing dynamic limits that
differentiate squad from non-squad connections.
- Rolled out targeted mechanisms for selecting optimal squad peers,
improving overall connection reliability.
  - Added functionality to specifically handle dialing non-squad peers.
- Introduced a method for classifying peers based on their status in
various lists.

- **Refactor**
- Overhauled the underlying connection management logic to adapt more
effectively to changing network conditions.
- Updated logging enhancements for improved tracking of connection
activities.
- Improved peer management with structured methods for classifying and
dialing peers.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->