fix: accept higher block tips and reduce catchup sync size (#331)
Change from ignoring higher new block notifies to accepting them, so
that the rest of the network receives them.
Also reduces the max number of blocks to send via catchup sync. This
will potentially slow down catchup sync, but currently there is a 10mb
limit on cbor for responses and I think we are hitting that limit.

In future it might be better to pass a number of blocks to sync so that
the client and server can negotiate a rate.

UPDATE: changed to 5 to be more conservative

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Reduced the maximum number of blocks returned during catch-up
synchronization from 10 to 5.
- Updated message handling to accept and propagate blocks significantly
ahead of the local tip instead of ignoring them.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->